+++ src_3_0_0/plugins/i18next.js
@@ -115,9 +115,10 @@ export function initLanguage (portalLanguageConfig) {
         });
     i18next.on("initialized", () => {
         if (!portalLanguage.enabled) {
-            i18next.changeLanguage("de");
+            i18next.changeLanguage("en");
         }
     });
     // @todo: remove if languages can be changed by language-tool, now you can use  i18next.changeLanguage("de"); on browser console
+    global.i18next = i18next;
     global.i18next = i18nextModule.i18next;
 }