+++ CHANGELOG.md
@@ -198,6 +198,7 @@
 ## v2.34.0 - 2023-06-07
 ### __Breaking Changes__
 - The Virtual City Planner tool (virtualcityPLANNER) has been removed
+- the `wfsSearch` modules config.json parameter 'field'-property `type` was renamed to `queryType`
 - new config.json parameter 'tree' added. Containes:
   - 'type' (was 'treeType' at root before), the following params are possible: "auto" ( = old "default") or "light"
   - 'validLayerTypesAutoTree' (new parameter) only for tree type 'auto'
@@ -378,7 +379,6 @@
         - @vue/test-utils: 1.3.0 to 2.2.3
         - mocha: 10.0.0 to 10.1.0
         - vue-loader: 15.9.8 to 17.0.0
-- in wfsSearch the 'field'-property type was renamed to queryType
 ### Removed
 - The following NPM packages are removed:
     - dependencies